# api-gateway/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/interviews": {
      "post": {
        "operationId": "InterviewsController_create",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interview scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Application or interviewer not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Schedule a new interview",
        "tags": [
          "interviews"
        ]
      },
      "get": {
        "operationId": "InterviewsController_findAll",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "applicationId",
            "required": false,
            "in": "query",
            "description": "Filter by application ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interviewerId",
            "required": false,
            "in": "query",
            "description": "Filter by interviewer ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PHONE",
                "VIDEO",
                "IN_PERSON",
                "PANEL",
                "TECHNICAL"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SCHEDULED",
                "CONFIRMED",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED",
                "NO_SHOW"
              ]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "scheduledAt",
              "type": "string",
              "enum": [
                "scheduledAt",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated interviews"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all interviews with pagination and filters",
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}": {
      "get": {
        "operationId": "InterviewsController_findOne",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Interview ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the interview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Interview not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an interview by ID",
        "tags": [
          "interviews"
        ]
      },
      "patch": {
        "operationId": "InterviewsController_update",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Interview ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interview updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Interview not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update/reschedule an interview",
        "tags": [
          "interviews"
        ]
      },
      "delete": {
        "operationId": "InterviewsController_remove",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Interview ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Interview cancelled successfully"
          },
          "404": {
            "description": "Interview not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cancel an interview",
        "tags": [
          "interviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "TalentFlow AI API",
    "description": "API documentation for the TalentFlow AI API Gateway",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your access token"
      }
    },
    "schemas": {
...
      "CreateInterviewDto": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "description": "Application ID"
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-03-20T10:00:00Z"
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes",
            "default": 60,
            "minimum": 15,
            "maximum": 480
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "VIDEO",
              "IN_PERSON",
              "PANEL",
              "TECHNICAL"
            ],
            "default": "VIDEO",
            "example": "VIDEO"
          },
          "location": {
            "type": "string",
            "example": "https://meet.google.com/abc-defg-hij"
          },
          "notes": {
            "type": "string",
            "example": "Focus on system design questions"
          },
          "interviewerId": {
            "type": "string",
            "description": "Interviewer user ID"
          }
        },
        "required": [
          "applicationId",
          "scheduledAt",
          "type"
        ]
      },
      "InterviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid"
          },
          "applicationId": {
            "type": "string",
            "example": "uuid"
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string"
          },
          "duration": {
            "type": "number",
            "example": 60
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "VIDEO",
              "IN_PERSON",
              "PANEL",
              "TECHNICAL"
            ]
          },
          "location": {
            "type": "object"
          },
          "notes": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "CONFIRMED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "interviewerId": {
            "type": "object",
            "example": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "applicationId",
          "scheduledAt",
          "duration",
          "type",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateInterviewDto": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "example": "2026-03-20T14:00:00Z"
          },
          "duration": {
            "type": "number",
            "description": "Duration in minutes",
            "minimum": 15,
            "maximum": 480
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "VIDEO",
              "IN_PERSON",
              "PANEL",
              "TECHNICAL"
            ]
          },
          "location": {
            "type": "string",
            "example": "https://zoom.us/j/123456"
          },
          "notes": {
            "type": "string",
            "example": "Candidate did well on system design"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "CONFIRMED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "interviewerId": {
            "type": "string",
            "description": "Interviewer user ID"
          }
        }
      },
...
    }
  }
}